LayoutTests/fast/scrolling/mac/unscrollable-root-overflow-hidden-on-one-axis-expected.txt

Tests that a 100% height root with hidden on one axis does not scroll.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS windowScrollEventCount is 0
PASS successfullyParsed is true

TEST COMPLETE
